Add vimrc and theme, update install script
The vim dir will is designed to be symlinked to .vim in ~/. The
install script will now check if any of the files already exist
in ~/, and if they do it will remove them and symlink the files
in the current working directory. If the config files do not exist
in ~/, it will create the symlinks.

+++ install.sh
@@ -2,6 +2,39 @@
 BASEDIR=$(dirname $0)
 cd $BASEDIR
 
-ln -s ${PWD}/bashrc ~/.bashrc
-ln -s ${PWD}/zshrc ~/.zshrc
-ln -s ${PWD}/tmux.conf ~/.tmux.conf
+
+if [ -e ~/.bashrc ];
+    then
+        rm ~/.bashrc && ln -s ${PWD}/bashrc ~/.bashrc
+    else
+	ln -s ${PWD}/bashrc ~/.bashrc
+fi
+
+if [ -e ~/.zshrc ];
+    then
+        rm ~/.zshrc && ln -s ${PWD}/zshrc ~/.zshrc
+    else
+	ln -s ${PWD}/zshrc ~/.zshrc
+
+fi
+
+if [ -e ~/.tmux.conf ];
+    then
+        rm ~/.tmux.conf && ln -s ${PWD}/tmux.conf ~/.tmux.conf
+    else
+	ln -s ${PWD}/tmux.conf ~/.tmux.conf
+fi
+
+if [ -d ~/.vim ];
+    then
+        rm -R ~/.vim && ln -s ${PWD}/vim/ ~/.vim
+    else
+        ln -s ${PWD}/vim/ ~/.vim
+fi
+
+if [ -e ~/.vimrc  ];
+    then
+        rm ~/.vimrc && ln -s ${PWD}/vimrc ~/.vimrc
+    else
+        ln -s ${PWD}/vimrc ~/.vimrc
+fi
